backend/requirements.txt
fastapi
uvicorn
groq
python-dotenv
faiss-cpu
sentence-transformers
numpy
pydantic
